refactor: 增加组扩展方法
