+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-extensions.js
@@ -1,16 +1,11 @@
-﻿import {
-    DockviewComponent,
-    DockviewGroupPanel,
-    getGridLocation,
-    getRelativeLocation,
-    DockviewEmitter
-} from "./dockview-core.esm.js"
+﻿import { DockviewComponent, DockviewGroupPanel, getGridLocation, getRelativeLocation, DockviewEmitter } from "./dockview-core.esm.js"
 import { getLocal } from './dockview-utils.js'
 
 DockviewComponent.prototype.on = function (eventType, callback) {
     this['_' + eventType] = new DockviewEmitter();
     this['_' + eventType].event(callback)
 }
+
 DockviewGroupPanel.prototype.getParams = function () {
     return this.activePanel?.params || {}
 }
@@ -27,10 +22,9 @@ DockviewGroupPanel.prototype.removePropsOfParams = function (keys) {
         : typeof keys == 'string' ? delete panel.params[keys] : false
 }
 
-// 修改removeGroup
 const removeGroup = DockviewComponent.prototype.removeGroup
 DockviewComponent.prototype.removeGroup = function (...argu) {
-    if(this.isClearIng){
+    if (this.isClearIng) {
         return removeGroup.apply(this, argu)
     }
 
@@ -56,18 +50,18 @@ DockviewComponent.prototype.removeGroup = function (...argu) {
         return removeGroup.apply(this, argu)
     }
 }
-// 修改removePanel
+
 const removePanel = DockviewComponent.prototype.removePanel
 DockviewComponent.prototype.removePanel = function (...argu) {
     const panel = argu[0]
     if (!panel.group.locked) {
         removePanel.apply(this, argu)
-        if(!this.isClearIng){
+        if (!this.isClearIng) {
             this._panelClosed?.fire(panel.title)
         }
     }
 }
-// 修改moveGroupOrPanel
+
 DockviewComponent.prototype.moveGroupOrPanel = function moveGroupOrPanel(options) {
     var _a;
     const destinationGroup = options.to.group;
@@ -174,14 +168,14 @@ DockviewComponent.prototype.moveGroupOrPanel = function moveGroupOrPanel(options
         }
     }
 }
-// moveGroupOrPanel源码提供1
+
 const tail = arr => {
     if (arr.length === 0) {
         throw new Error('Invalid tail call');
     }
     return [arr.slice(0, arr.length - 1), arr[arr.length - 1]];
 }
-// moveGroupOrPanel源码提供2
+
 const sequenceEquals = (arr1, arr2) => {
     if (arr1.length !== arr2.length) {
         return false;
